Relocate Framework Attributes for 4.x
