+++ src/Phaseolies/Console/Commands/stubs/auth/HomeController.stub
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers;
 
 use Phaseolies\Http\Response;
-use Phaseolies\Utilities\Attributes\Route;
-use Phaseolies\Utilities\Attributes\Mapper;
+use Phaseolies\Support\Router\Attributes\Route;
+use Phaseolies\Support\Router\Attributes\Mapper;
 use App\Http\Controllers\Controller;
 
 #[Mapper(prefix: 'home', middleware: ['auth'])]
