+++ src/Phaseolies/Console/Commands/stubs/auth/RegisterController.stub
@@ -5,8 +5,8 @@ namespace App\Http\Controllers\Auth;
 use Phaseolies\Http\Request;
 use Phaseolies\Http\Response;
 use Phaseolies\Http\Response\RedirectResponse;
-use Phaseolies\Utilities\Attributes\Route;
-use Phaseolies\Utilities\Attributes\Mapper;
+use Phaseolies\Support\Router\Attributes\Route;
+use Phaseolies\Support\Router\Attributes\Mapper;
 use App\Models\User;
 use App\Http\Controllers\Controller;
