+++ src/Phaseolies/Console/Commands/stubs/controller/complete.stub
@@ -3,7 +3,7 @@
 namespace {{ namespace }};
 
 use App\Http\Controllers\Controller;
-use Phaseolies\Utilities\Attributes\Route;
+use Phaseolies\Support\Router\Attributes\Route;
 
 class {{ class }} extends Controller
 {
