+++ src/Phaseolies/Middleware/Attributes/Middleware.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phaseolies\Utilities\Attributes;
+namespace Phaseolies\Middleware\Attributes;
 
 use Attribute;
